Setup updates, fixes and improvements
- Updates browser-sync and engine-phtml
- Renames and adds new npm and JS scripts
